refactor(mv): split src/core/mv.rs into responsibility-oriented submodules
Break the 1750-line src/core/mv.rs into a module directory organized
by responsibility, without any behavior or public API change:

- validate.rs: path resolution and existence/collision validation
- case_only.rs: case-only rename detection and planning
- plan.rs: replacement planning (external/internal/directory), link
  rewriting construction, ReplacementPlan bookkeeping
- apply.rs: filesystem mutation (apply_replacements, line-ending
  preservation, rename-with-cross-device-fallback, rollback execution)
- preview.rs: build_move_preview and print_dry_run_report
- mod.rs: pub fn mv / pub fn preview_move and the three orchestration
  routines (regular file / case-only / directory), plus the existing
  tests module (kept intact to preserve coverage and minimize risk)

The MoveTransaction data model already lives in
src/core/model/move_transaction.rs and is reused as-is.

All 275 tests pass; scripts/precheck.sh passes locally.

Refs #10
